moved functionality to library
