+++ minigrep/src/lib.rs
@@ -1,0 +1,31 @@
+use std::{error::Error, fs};
+
+pub struct Config {
+    query: String,
+    file_path: String,
+}
+
+impl Config {
+    pub fn build(args: &[String]) -> Result<Config, &'static str> {
+        if args.len() < 3 {
+            return Err("not enough arguments");
+        }
+        let query = args[1].clone();
+        let file_path = args[2].clone();
+
+        Ok(Config { query, file_path })
+    }
+    pub fn query(&self) -> &String {
+        &self.query
+    }
+
+    pub fn file_path(&self) -> &String {
+        &self.file_path
+    }
+}
+
+pub fn run(config: Config) -> Result<(), Box<dyn Error>> {
+    let contents = fs::read_to_string(config.file_path)?;
+    println!("With text:\n{contents}");
+    Ok(())
+}
