+++ minigrep/src/main.rs
@@ -1,4 +1,6 @@
-use std::{env, error::Error, fs, process};
+use std::{env, process};
+
+use minigrep::{run, Config};
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -6,33 +8,10 @@ fn main() {
         println!("Problem parsing arguments: {err}");
         process::exit(1);
     });
-    println!("Searching for '{}'", config.query);
-    println!("In file '{}'", config.file_path);
+    println!("Searching for '{}'", config.query());
+    println!("In file '{}'", config.file_path());
     if let Err(e) = run(config) {
         println!("Application errored reading file: {e}");
         process::exit(1);
     }
-}
-
-struct Config {
-    query: String,
-    file_path: String,
-}
-
-impl Config {
-    fn build(args: &[String]) -> Result<Config, &'static str> {
-        if args.len() < 3 {
-            return Err("not enough arguments");
-        }
-        let query = args[1].clone();
-        let file_path = args[2].clone();
-
-        Ok(Config { query, file_path })
-    }
-}
-
-fn run(config: Config) -> Result<(), Box<dyn Error>> {
-    let contents = fs::read_to_string(config.file_path)?;
-    println!("With text:\n{contents}");
-    Ok(())
 }
